+++ pkg/network/message/types.go
@@ -9,7 +9,6 @@ type IMessage interface {
 	types.IConverter
 
 	GetHash() []byte
-	GetRand() []byte
 	GetProof() uint64
 
 	// payload = head(32bit) || body(Nbit)
@@ -19,7 +18,6 @@ type IMessage interface {
 type IConstructSettings interface {
 	GetSettings() ISettings
 	GetParallel() uint64
-	GetRandMessageSizeBytes() uint64
 }
 
 type ISettings interface {