+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### CHANGES
 
 - `pkg/crypto/hashing`: doOnce encoding.HexEncode -> ToString method
+- `pkg/network/message`: delete GetRand, GetRandMessageSizeBytes
 
 <!-- ... -->